backend/requirements.txt
Django==2.2
djangorestframework==3.9.4
gpxpy==1.3.5
